Added package/package-lock.json, Setup Login/Signup page
Modified page's logged state as routes only accessible from the authenticated view will be set to true with the user. Pushed package.json and package-lock.json as a test for other team members to not have to install every dependency that is not included in server/front side. Forms in the Login/SigupView needed CSS styling